+++ CHANGELOG.md
@@ -1,10 +1,24 @@
+## v0.0.5
+
+* Add: server and client now support macOS.
+* Add: shutdown and reboot on macOS (root permission required).
+* Update: pty are used on Unix-like OS to provide a full-functional terminal.
+* Update: improved the support of terminal on Windows and fixed some bugs.
+
+* 新增：服务端和客户端已支持macOS系统。
+* 新增：macOS现在将支持关机和重启功能（需要root权限）。
+* 更新：类unix系统的终端现已改用pty实现，以提供完整的终端功能。
+* 更新：改进了windows下的终端表现，修复了一些bug。
+
+
+
 ## v0.0.4
 
-* Add: i18n support
+* Add: i18n support.
 * Note: from this version, you just need to upgrade server manually and client will automatically self upgrade.
 
-* 新增：支持中英文国际化
-* 注意：从本版本开始，只需要更新服务端即可，客户端会自动完成更新
+* 新增：支持中英文国际化。
+* 注意：从本版本开始，只需要更新服务端即可，客户端会自动完成更新。